# Conflicts:
#	TICC/src/main/java/uy/edu/um/bbticg4/service/entities/Restaurant.java

+++ TICC/src/main/java/uy/edu/um/bbticg4/service/ReservaMgr.java
@@ -28,7 +28,7 @@ public class ReservaMgr {
     @Autowired
     private MesaRepository mesaRepo;
 
-    public void generarReserva(ClienteFinal cf, Restaurant resto, Integer cantidad, LocalDateTime hora)
+    public void generarReserva(ClienteFinal cf, Restaurant resto, Double cantidad, LocalDateTime hora)
             throws InvalidInformation {
         if(cf == null || resto == null || cantidad == null || hora == null ||
                 !cfr.existsById(cf.getId()) || !restoRepo.existsById(resto.getId()) ||
